+++ blueprints/base.py
@@ -1,0 +1,4 @@
+from flask import Flask
+
+
+app = Flask(__name__, template_folder='templates', static_folder="../static")
